requirements.txt
jira>=3.5.0
python-dotenv>=1.0.0
requests>=2.28.0
flask>=2.0.0
openai>=1.0.0
google-genai>=1.0.0
gspread>=6.0.0
google-auth>=2.0.0